Relocate AAP Client and Provisioning Package
The bare-metal-fulfillment-operator wants to reuse the code in this
package for its way of setting up/tearing down clusters.

Closes osac-project/issues#387
