+++ internal/controller/computeinstance_provisioning_test.go
@@ -28,7 +28,7 @@ import (
 	mcmanager "sigs.k8s.io/multicluster-runtime/pkg/manager"
 
 	osacv1alpha1 "github.com/osac-project/osac-operator/api/v1alpha1"
-	"github.com/osac-project/osac-operator/internal/provisioning"
+	"github.com/osac-project/osac-operator/pkg/provisioning"
 )
 
 // mockProvisioningProvider implements the ProvisioningProvider interface for testing
