+++ pkg/aap/client_test.go
@@ -13,7 +13,7 @@ import (
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 
-	"github.com/osac-project/osac-operator/internal/aap"
+	"github.com/osac-project/osac-operator/pkg/aap"
 )
 
 func TestAAP(t *testing.T) {
