+++ pkg/provisioning/aap_provider.go
@@ -12,7 +12,7 @@ import (
 	ctrllog "sigs.k8s.io/controller-runtime/pkg/log"
 
 	"github.com/osac-project/osac-operator/api/v1alpha1"
-	"github.com/osac-project/osac-operator/internal/aap"
+	"github.com/osac-project/osac-operator/pkg/aap"
 )
 
 // AAPClient is the interface for AAP operations used by the provider.
